Module rolling_stats::partial_data_buffer

PartialDataBuffer represents a way of dealing with incomplete raw data. It consumes a slice of the newly received data, saves the data that are required for completing a previously incomplete data, parses them and returns a new slice, that contains valid data for further processing. Trailing incomplete data are handled as well. The new returned slice has appropriate size so that an integer number of values can be parsed using it.

As opposed to the Reconstructor, this solution avoids pointless copies.

Structs

PartialDataBuffer
Implements the partial data buffer - handling of incomplete data in a stream of raw data.