Trait rolling_stats::ConverterFromRaw

pub trait ConverterFromRaw<T> {
    fn from_raw(raw: &[u8]) -> Result<T, RawConversionError>;
}

Trait utilized for implementing conversion of raw bytes into specific types. Implemented by Converter structs such as the LittleEndian and BigEndian structs. T denotes the type the raw bytes should be converted into.

Required methods

fn from_raw(raw: &[u8]) -> Result<T, RawConversionError>

Returns either the converted type from the raw input or an error.

Arguments

raw - raw bytes the type will be reconstructed from, length should be the same or longer than the type itself.

Loading content...