Trait rolling_stats::Statistics

pub trait Statistics {
    fn mean(&self) -> f32;

    fn std_dev(&self) -> f32;

    fn rand(&self) -> f32;
}

The Statistics trait useful for dependency injection. This trait abstracts away basic statistics measures.

Required methods

fn mean(&self) -> f32

Returns the mean of a dataset.

fn std_dev(&self) -> f32

Returns standard deviation of a dataset.

fn rand(&self) -> f32

Returns a number from a standard distribution specified by the mean and standard deviation of the dataset.

Loading content...